nonebot_plugin_talk_with_poe_ai/pyproject.toml
[tool.poetry]
name = "nonebot_plugin_talk_with_poe_ai"
version = "0.1.0"
description = "Nonebot2 基于poe cookie登录AI聊天插件"
authors = ["nikissXI <1299577815@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_talk_with_poe_ai"}]
homepage = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_talk_with_poe_ai"
repository = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_talk_with_poe_ai"
documentation = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_talk_with_poe_ai#README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = ">=3.8"
nonebot2 = ">=2.0.0rc1"
nonebot-adapter-onebot = ">=2.0.0"
aiohttp = ">=3.8.4"
aiohttp_socks = ">=0.8.0"
Pillow = ">=9.0.0"
loguru = ">=0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
